# lib/api-zod/src/generated/types/campaignSendInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ditheto Accountants administration API
 * OpenAPI spec version: 0.2.0
 */

export interface CampaignSendInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  /** @minLength 1 */
  objectPath?: string;
}