# lib/api-zod/src/generated/types/campaignSendResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ditheto Accountants administration API
 * OpenAPI spec version: 0.2.0
 */
import type { CampaignSendResultChannel } from './campaignSendResultChannel';
import type { CampaignSendResultStatus } from './campaignSendResultStatus';

export interface CampaignSendResult {
  success: boolean;
  status: CampaignSendResultStatus;
  channel: CampaignSendResultChannel;
  deliveryUrl: string;
  message: string;
}